# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Product:
    properties:
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  http.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  product.ProductListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      products:
        items:
          $ref: '#/definitions/entity.Product'
        type: array
      total:
        type: integer
    type: object
  product.ProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample e-commerce API server.
  title: E-commerce API
  version: "1.0"
paths:
...
  /api/v1/admin/export/products:
    post:
      consumes:
      - application/json
      description: Download all products as CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema: {}
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export products to CSV (Admin only)
      tags:
      - Admin - Export
...
  /api/v1/admin/products:
    post:
      consumes:
      - application/json
      description: Create a new product with details. Requires admin role.
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product (Admin only)
      tags:
      - Admin - Products
  /api/v1/admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by its ID. Requires admin role.
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/http.SuccessResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product (Admin only)
      tags:
      - Admin - Products
    put:
      consumes:
      - application/json
      description: Update details of an existing product by its ID. Requires admin
        role.
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Invalid input or product ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing product (Admin only)
      tags:
      - Admin - Products
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with optional filtering and
        searching
      parameters:
      - description: Search term for product name or description
        in: query
        name: search
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - default: 20
        description: Number of items to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            $ref: '#/definitions/product.ProductListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get a list of products
      tags:
      - Products
  /api/v1/products/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a single product by its ID
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/entity.Product'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get product by ID
      tags:
      - Products
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"